+++ tests/strands/tools/test_loader.py
@@ -1,136 +1,12 @@
 import os
-import pathlib
 import re
 import textwrap
-import unittest.mock
 
 import pytest
 
-import strands
 from strands.tools.decorator import DecoratedFunctionTool
 from strands.tools.loader import ToolLoader
 from strands.tools.tools import PythonAgentTool
-
-
-def test_load_function_tool():
-    @strands.tools.tool
-    def tool_function(a):
-        return a
-
-    tool = strands.tools.loader.load_function_tool(tool_function)
-
-    assert isinstance(tool, DecoratedFunctionTool)
-
-
-def test_load_function_tool_no_function():
-    tool = strands.tools.loader.load_function_tool("no_function")
-
-    assert tool is None
-
-
-def test_load_function_tool_no_spec():
-    def tool_function(a):
-        return a
-
-    tool = strands.tools.loader.load_function_tool(tool_function)
-
-    assert tool is None
-
-
-def test_load_function_tool_invalid():
-    def tool_function(a):
-        return a
-
-    tool_function.TOOL_SPEC = "invalid"
-
-    tool = strands.tools.loader.load_function_tool(tool_function)
-
-    assert tool is None
-
-
-def test_scan_module_for_tools():
-    @strands.tools.tool
-    def tool_function_1(a):
-        return a
-
-    @strands.tools.tool
-    def tool_function_2(b):
-        return b
-
-    def tool_function_3(c):
-        return c
-
-    def tool_function_4(d):
-        return d
-
-    tool_function_4.tool_spec = "invalid"
-
-    mock_module = unittest.mock.MagicMock()
-    mock_module.tool_function_1 = tool_function_1
-    mock_module.tool_function_2 = tool_function_2
-    mock_module.tool_function_3 = tool_function_3
-    mock_module.tool_function_4 = tool_function_4
-
-    tools = strands.tools.loader.scan_module_for_tools(mock_module)
-
-    assert len(tools) == 2
-    assert all(isinstance(tool, DecoratedFunctionTool) for tool in tools)
-
-
-def test_scan_directory_for_tools(tmp_path):
-    tool_definition_1 = textwrap.dedent("""
-        import strands
-
-        @strands.tools.tool
-        def tool_function_1(a):
-            return a
-    """)
-    tool_definition_2 = textwrap.dedent("""
-        import strands
-
-        @strands.tools.tool
-        def tool_function_2(b):
-            return b
-    """)
-    tool_definition_3 = textwrap.dedent("""
-        def tool_function_3(c):
-            return c
-    """)
-    tool_definition_4 = textwrap.dedent("""
-        def tool_function_4(d):
-            return d
-    """)
-    tool_definition_5 = ""
-    tool_definition_6 = "**invalid**"
-
-    tool_path_1 = tmp_path / "tool_1.py"
-    tool_path_2 = tmp_path / "tool_2.py"
-    tool_path_3 = tmp_path / "tool_3.py"
-    tool_path_4 = tmp_path / "tool_4.py"
-    tool_path_5 = tmp_path / "_tool_5.py"
-    tool_path_6 = tmp_path / "tool_6.py"
-
-    tool_path_1.write_text(tool_definition_1)
-    tool_path_2.write_text(tool_definition_2)
-    tool_path_3.write_text(tool_definition_3)
-    tool_path_4.write_text(tool_definition_4)
-    tool_path_5.write_text(tool_definition_5)
-    tool_path_6.write_text(tool_definition_6)
-
-    tools = strands.tools.loader.scan_directory_for_tools(tmp_path)
-
-    tru_tool_names = sorted(tools.keys())
-    exp_tool_names = ["tool_function_1", "tool_function_2"]
-
-    assert tru_tool_names == exp_tool_names
-    assert all(isinstance(tool, DecoratedFunctionTool) for tool in tools.values())
-
-
-def test_scan_directory_for_tools_does_not_exist():
-    tru_tools = strands.tools.loader.scan_directory_for_tools(pathlib.Path("does_not_exist"))
-    exp_tools = {}
-
-    assert tru_tools == exp_tools
 
 
 @pytest.fixture
