+++ tests/strands/tools/test_registry.py
@@ -7,6 +7,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from strands.tools import PythonAgentTool
+from strands.tools.decorator import DecoratedFunctionTool, tool
 from strands.tools.registry import ToolRegistry
 
 
@@ -43,3 +44,34 @@ def test_register_tool_with_similar_name_raises():
         str(err.value) == "Tool name 'tool_like_this' already exists as 'tool-like-this'. "
         "Cannot add a duplicate tool which differs by a '-' or '_'"
     )
+
+
+def test_scan_module_for_tools():
+    @tool
+    def tool_function_1(a):
+        return a
+
+    @tool
+    def tool_function_2(b):
+        return b
+
+    def tool_function_3(c):
+        return c
+
+    def tool_function_4(d):
+        return d
+
+    tool_function_4.tool_spec = "invalid"
+
+    mock_module = MagicMock()
+    mock_module.tool_function_1 = tool_function_1
+    mock_module.tool_function_2 = tool_function_2
+    mock_module.tool_function_3 = tool_function_3
+    mock_module.tool_function_4 = tool_function_4
+
+    tool_registry = ToolRegistry()
+
+    tools = tool_registry._scan_module_for_tools(mock_module)
+
+    assert len(tools) == 2
+    assert all(isinstance(tool, DecoratedFunctionTool) for tool in tools)
